Travis: only build branch builds on master
Travis has two types of builds:

 1. Pull Request builds get run on branches other than master
 2. Branch builds run on any push to any branch

We enable both in settings, but only run the latter on master. This makes
it so pull requests get pull request builds and master gets a branch build
(on merge), but branches don't get _two_ builds (a Pull Request build and
a Branch build).